requirements.txt
google-adk>=1.2.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
python-dotenv>=1.0.0
websockets>=14.0
pillow>=11.0.0
